+++ Sources/SwiftCSP/Constraint.swift
@@ -24,7 +24,7 @@ open class Constraint <V: Hashable, D> {
     ///
     /// - parameter assignment: Potential domain selections for variables that are part of the constraint.
     /// - returns: Whether the constraint is satisfied.
-    public func isSatisfied(assignment: Dictionary<V, D>) -> Bool {
+    open func isSatisfied(assignment: Dictionary<V, D>) -> Bool {
         return true
     }
     /// The variables that make up the constraint.